fix warnings

+++ Classes/WEPopoverTableViewController.m
@@ -106,7 +106,7 @@
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
     }
 	
-	cell.textLabel.text = [NSString stringWithFormat:@"Cell %d", indexPath.row];
+	cell.textLabel.text = [NSString stringWithFormat:@"Cell %ld", (long)indexPath.row];
     
     return cell;
 }
